Rename BreadcrumbsLogger to Breadcrumbs::SentryLogger
